+++ test_lib/constrained_jungle.ex
@@ -3,10 +3,10 @@ defmodule ConstrainedJungle do
 
   open_api_object "ConstrainedJungle", deprecated: true do
     property :page, :integer, minimum: 1, maximum: 100
-    property :slug, :string, minLength: 3, maxLength: 30
-    property :tags, {:array, :string}, minItems: 1, example: ["red", "blue"]
-    property :ids, {:array, :string}, format: :uuid, minItems: 2
-    property :codes, {:map, :string}, minLength: 2, minProperties: 1, example: %{"a" => "bb"}
+    property :slug, :string, min_length: 3, max_length: 30
+    property :tags, {:array, :string}, min_items: 1, example: ["red", "blue"]
+    property :ids, {:array, :string}, format: :uuid, min_items: 2
+    property :codes, {:map, :string}, min_length: 2, min_properties: 1, example: %{"a" => "bb"}
     property :either, [:integer, :string], description: "a number or a word", example: 5
     property :rank, :integer, enum: [1, 2, 3]
   end